Guest Functionality
Main functionality of having guest orders complete!

Based on if ur a guest or a customer, the system processes ur order differently. Need to make sure info is being pulled from the checkout page - make it so it only asks name etc if you're not logged in becvause its a bit repetitivbe